Add check if bucket is locked to not remove tabs.

+++ src/scripts/internal/tab-signals.ts
@@ -100,7 +100,9 @@ export class TabSignals {
 
       if (archiveBucket) {
         archiveBucket.removeTabs();
-        window.globalBucketTabsState.BucketListDataModel.saveArchiveBucket(archiveBucket);
+        window.globalBucketTabsState.BucketListDataModel.saveArchiveBucket(
+          archiveBucket
+        );
 
         // View
         window.globalBucketTabsState.BucketSignals.emit(
@@ -116,32 +118,39 @@ export class TabSignals {
       window.globalBucketTabsState.BucketListDataModel.getBucketByID(bucketID);
 
     if (bucket) {
-
       bucket.getTabs().forEach((tab) => {
-        console.log("TODO - Restore tab: ", tab.getTabURL());
+        console.log('TODO - Restore tab: ', tab.getTabURL());
       });
 
-      bucket.removeTabs();
-      window.globalBucketTabsState.BucketListDataModel.saveBucket(bucket);
+      if (!bucket.getLocked()) {
+        bucket.removeTabs();
+        window.globalBucketTabsState.BucketListDataModel.saveBucket(bucket);
 
-      // View
-      window.globalBucketTabsState.BucketSignals.emit(BucketEvents.LoadBuckets);
+        // View
+        window.globalBucketTabsState.BucketSignals.emit(
+          BucketEvents.LoadBuckets
+        );
+      }
     } else {
       let archiveBucket =
         window.globalBucketTabsState.BucketListDataModel.getArchivedBuckets()[0];
 
       if (archiveBucket) {
         archiveBucket.getTabs().forEach((tab) => {
-          console.log("TODO - Restore tab: ", tab.getTabURL());
+          console.log('TODO - Restore tab: ', tab.getTabURL());
         });
-        
-        archiveBucket.removeTabs();
-        window.globalBucketTabsState.BucketListDataModel.saveArchiveBucket(archiveBucket);
 
-        // View
-        window.globalBucketTabsState.BucketSignals.emit(
-          BucketEvents.LoadArchive
-        );
+        if (!archiveBucket.getLocked()) {
+          archiveBucket.removeTabs();
+          window.globalBucketTabsState.BucketListDataModel.saveArchiveBucket(
+            archiveBucket
+          );
+
+          // View
+          window.globalBucketTabsState.BucketSignals.emit(
+            BucketEvents.LoadArchive
+          );
+        }
       }
     }
   }
